Update HostPinger to be immutable
- MerlinService.hostPinger was final but it made more sense for
HostPinger to be immutable, and make the MerlinService.hostPinger
non-final
- This should not affect external usage because MerlinService.hostPinger
was private and only created internal to MerlinService
- Making the HostPinger immutable made it simpler - host name cannot
change, so no need to test these

This commit introduces a test coverage regression - the callbacks are
not tested from HostPinger anymore. This is because the callbacks are
used in an AsyncTask, which on Android Unit Tests is just a set of empty
code blocks.

The logic from PingTask was extracted into Ping (synchronous), but the
result could not be.